Enable waveforms > 2^15 samples (res short => int) On branch skope 	modified:   padwaveforms.m 	modified:   readmod.m 	modified:   writemod.m

+++ +toppe/readmod.m
@@ -41,7 +41,7 @@ desc = char(desc');
 
 % read rest of header
 hdr.ncoils = fread(fid, 1, 'int16');
-hdr.res    = fread(fid, 1, 'int16');
+hdr.res    = fread(fid, 1, 'int32');
 hdr.npulses = fread(fid, 1, 'int16');
 hdr.b1max  = fscanf(fid, 'b1max:  %f\n');
 hdr.gmax   = fscanf(fid, 'gmax:   %f\n');
